make score changing
