+++ common.h
@@ -1,7 +1,7 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-#define VERSION "v2.4.1"
+const char* version = (char*)"v2.5.1";
 
 #include <cstdio>
 #include <cstdlib>
@@ -13,13 +13,6 @@
 #include <unistd.h>
 #include <termios.h>
 
-#ifndef false
-#define false 0
-#endif
-
-#ifndef true
-#define true 1
-#endif
 /* some special characters(terminal graph):
  *
  *  ▁ ▂ ▃ ▄ ▅ ▆ ▇ █ ▊ ▌ ▎ ▖ ▗ ▘ ▙ ▚ ▛ ▜ ▝ ▞ ▟ ━ ┃
@@ -164,12 +157,13 @@ typedef struct server_message_t {
         };
 
         struct {
-            char user_name[USER_CNT][USERNAME_SIZE];
-            uint8_t user_namecolor[USER_CNT];
-            uint8_t user_kill[USER_CNT];
-            uint8_t user_death[USER_CNT];
-            uint16_t user_life[USER_CNT];
-        };
+            char name[USERNAME_SIZE];
+            uint8_t namecolor;
+            uint8_t kill;
+            uint8_t death;
+            uint8_t score;
+            uint16_t life;
+        } users[USER_CNT];
 
         struct {
             char from_user[USERNAME_SIZE];
@@ -178,5 +172,85 @@ typedef struct server_message_t {
     };
 } server_message_t;
 
+enum {
+    CLIENT_COMMAND_USER_QUIT,
+    CLIENT_COMMAND_USER_REGISTER,
+    CLIENT_COMMAND_USER_LOGIN,
+    CLIENT_COMMAND_USER_LOGOUT,
+    CLIENT_MESSAGE_FATAL,
+    CLIENT_COMMAND_FETCH_ALL_USERS,
+    CLIENT_COMMAND_FETCH_ALL_FRIENDS,
+    CLIENT_COMMAND_LAUNCH_BATTLE,
+    CLIENT_COMMAND_LAUNCH_FFA,
+    CLIENT_COMMAND_QUIT_BATTLE,
+    CLIENT_COMMAND_ACCEPT_BATTLE,
+    CLIENT_COMMAND_REJECT_BATTLE,
+    CLIENT_COMMAND_INVITE_USER,
+    CLIENT_COMMAND_SEND_MESSAGE,
+    CLIENT_COMMAND_MOVE_UP,
+    CLIENT_COMMAND_MOVE_DOWN,
+    CLIENT_COMMAND_MOVE_LEFT,
+    CLIENT_COMMAND_MOVE_RIGHT,
+    CLIENT_COMMAND_FIRE_UP,
+    CLIENT_COMMAND_FIRE_DOWN,
+    CLIENT_COMMAND_FIRE_LEFT,
+    CLIENT_COMMAND_FIRE_RIGHT,
+    CLIENT_COMMAND_FIRE_UP_LEFT,
+    CLIENT_COMMAND_FIRE_UP_RIGHT,
+    CLIENT_COMMAND_FIRE_DOWN_LEFT,
+    CLIENT_COMMAND_FIRE_DOWN_RIGHT,
+    CLIENT_COMMAND_FIRE_AOE_UP,
+    CLIENT_COMMAND_FIRE_AOE_DOWN,
+    CLIENT_COMMAND_FIRE_AOE_LEFT,
+    CLIENT_COMMAND_FIRE_AOE_RIGHT,
+    CLIENT_COMMAND_ADMIN_CONTROL,
+    CLIENT_COMMAND_END,
+};
+
+enum {
+    SERVER_SAY_NOTHING,
+    SERVER_RESPONSE_REGISTER_SUCCESS,
+    SERVER_RESPONSE_REGISTER_FAIL,
+    SERVER_RESPONSE_YOU_HAVE_REGISTERED,
+    SERVER_RESPONSE_LOGIN_SUCCESS,
+    SERVER_RESPONSE_YOU_HAVE_LOGINED,
+    SERVER_RESPONSE_YOU_HAVE_NOT_LOGIN,
+    SERVER_RESPONSE_LOGIN_FAIL_UNREGISTERED_USERID,
+    SERVER_RESPONSE_LOGIN_FAIL_ERROR_PASSWORD,
+    SERVER_RESPONSE_LOGIN_FAIL_DUP_USERID,    // user id has been registered by other users
+    SERVER_RESPONSE_LOGIN_FAIL_SERVER_LIMITS, // server unable to handle more users
+    SERVER_RESPONSE_ALL_USERS_INFO,
+    SERVER_RESPONSE_ALL_FRIENDS_INFO,
+    SERVER_RESPONSE_LAUNCH_BATTLE_FAIL,
+    SERVER_RESPONSE_LAUNCH_BATTLE_SUCCESS,
+    SERVER_RESPONSE_YOURE_NOT_IN_BATTLE,
+    SERVER_RESPONSE_YOURE_ALREADY_IN_BATTLE,
+    SERVER_RESPONSE_INVITATION_SENT,
+    SERVER_RESPONSE_NOBODY_INVITE_YOU,
+    /* ----------------------------------------------- */
+    SERVER_MESSAGE,
+    SERVER_MESSAGE_DELIM,
+    SERVER_MESSAGE_FRIEND_LOGIN,
+    SERVER_MESSAGE_FRIEND_LOGOUT,
+    SERVER_MESSAGE_FRIEND_ACCEPT_BATTLE,
+    SERVER_MESSAGE_FRIEND_REJECT_BATTLE,
+    SERVER_MESSAGE_FRIEND_NOT_LOGIN,
+    SERVER_MESSAGE_FRIEND_ALREADY_IN_BATTLE,
+    SERVER_MESSAGE_INVITE_TO_BATTLE,
+    SERVER_MESSAGE_FRIEND_MESSAGE,
+    SERVER_MESSAGE_USER_QUIT_BATTLE,
+    SERVER_MESSAGE_BATTLE_DISBANDED,          // since no other users in this battle
+    SERVER_MESSAGE_BATTLE_INFORMATION,
+    SERVER_MESSAGE_BATTLE_PLAYER,
+    SERVER_MESSAGE_YOU_ARE_DEAD,
+    SERVER_MESSAGE_YOU_ARE_SHOOTED,
+    SERVER_MESSAGE_YOU_ARE_TRAPPED_IN_MAGMA,
+    SERVER_MESSAGE_YOU_GOT_BLOOD_VIAL,
+    SERVER_MESSAGE_YOU_GOT_MAGAZINE,
+    SERVER_MESSAGE_YOUR_MAGAZINE_IS_EMPTY,
+
+    SERVER_STATUS_QUIT,
+    SERVER_STATUS_FATAL,
+};
 
 #endif
