+++ constants.h
@@ -14,7 +14,7 @@
 
 #define IPADDR_SIZE 24
 #define USERNAME_SIZE  12
-#define MSG_SIZE 40
+#define MSG_SIZE 96
 #define USER_CNT   15
 
 #define PASSWORD_SIZE USERNAME_SIZE
@@ -34,7 +34,47 @@
 
 #define BULLETS_LASTS_TIME 300
 #define OTHER_ITEM_LASTS_TIME 1000
-#define GLOBAL_SPEED 30
+#define GLOBAL_SPEED 20
+
+enum {
+    ITEM_NONE,
+    ITEM_MAGAZINE,
+    ITEM_MAGMA,
+    ITEM_GRASS,
+    ITEM_BLOOD_VIAL,
+    ITEM_END,
+    ITEM_BULLET,
+};
+
+enum {
+    MAP_ITEM_NONE,
+    MAP_ITEM_MY_BULLET,
+    MAP_ITEM_OTHER_BULLET,
+    MAP_ITEM_MAGAZINE,
+    MAP_ITEM_BLOOD_VIAL,
+    MAP_ITEM_MAGMA,
+    MAP_ITEM_GRASS,
+    MAP_ITEM_USER,
+    MAP_ITEM_END,
+};
+
+enum {
+    DIR_UP,
+    DIR_DOWN,
+    DIR_LEFT,
+    DIR_RIGHT,
+    DIR_UP_LEFT,
+    DIR_UP_RIGHT,
+    DIR_DOWN_LEFT,
+    DIR_DOWN_RIGHT,
+};
+
+enum {
+    BATTLE_STATE_UNJOINED,
+    BATTLE_STATE_LIVE,
+    BATTLE_STATE_WITNESS,
+    BATTLE_STATE_DEAD,
+};
 
 /* unused  -->  not login  -->  login  <-...->  battle
  *
@@ -78,28 +118,6 @@ char* color_s[128];
 char*  map_s[128];
 int item_to_map[128];
 int color_s_size;
-
-enum {
-    ITEM_NONE,
-    ITEM_MAGAZINE,
-    ITEM_MAGMA,
-    ITEM_GRASS,
-    ITEM_BLOOD_VIAL,
-    ITEM_END,
-    ITEM_BULLET,
-};
-
-enum {
-    MAP_ITEM_NONE,
-    MAP_ITEM_MAGAZINE,
-    MAP_ITEM_BLOOD_VIAL,
-    MAP_ITEM_MAGMA,
-    MAP_ITEM_MY_BULLET,
-    MAP_ITEM_OTHER_BULLET,
-    MAP_ITEM_GRASS,
-    MAP_ITEM_USER,
-    MAP_ITEM_END,
-};
 
 void init_constants() {
     map_s[MAP_ITEM_NONE] = (char*)" ";
@@ -163,103 +181,4 @@ void init_constants() {
     color_s_size = 12;
 };
 
-enum {
-    CLIENT_COMMAND_USER_QUIT,
-    CLIENT_COMMAND_USER_REGISTER,
-    CLIENT_COMMAND_USER_LOGIN,
-    CLIENT_COMMAND_USER_LOGOUT,
-    CLIENT_MESSAGE_FATAL,
-    CLIENT_COMMAND_FETCH_ALL_USERS,
-    CLIENT_COMMAND_FETCH_ALL_FRIENDS,
-    CLIENT_COMMAND_LAUNCH_BATTLE,
-    CLIENT_COMMAND_LAUNCH_FFA,
-    CLIENT_COMMAND_QUIT_BATTLE,
-    CLIENT_COMMAND_ACCEPT_BATTLE,
-    CLIENT_COMMAND_REJECT_BATTLE,
-    CLIENT_COMMAND_INVITE_USER,
-    CLIENT_COMMAND_SEND_MESSAGE,
-    CLIENT_COMMAND_MOVE_UP,
-    CLIENT_COMMAND_MOVE_DOWN,
-    CLIENT_COMMAND_MOVE_LEFT,
-    CLIENT_COMMAND_MOVE_RIGHT,
-    CLIENT_COMMAND_FIRE_UP,
-    CLIENT_COMMAND_FIRE_DOWN,
-    CLIENT_COMMAND_FIRE_LEFT,
-    CLIENT_COMMAND_FIRE_RIGHT,
-    CLIENT_COMMAND_FIRE_UP_LEFT,
-    CLIENT_COMMAND_FIRE_UP_RIGHT,
-    CLIENT_COMMAND_FIRE_DOWN_LEFT,
-    CLIENT_COMMAND_FIRE_DOWN_RIGHT,
-    CLIENT_COMMAND_FIRE_AOE_UP,
-    CLIENT_COMMAND_FIRE_AOE_DOWN,
-    CLIENT_COMMAND_FIRE_AOE_LEFT,
-    CLIENT_COMMAND_FIRE_AOE_RIGHT,
-    CLIENT_COMMAND_ADMIN_CONTROL,
-    CLIENT_COMMAND_END,
-};
-
-enum {
-    SERVER_SAY_NOTHING,
-    SERVER_RESPONSE_REGISTER_SUCCESS,
-    SERVER_RESPONSE_REGISTER_FAIL,
-    SERVER_RESPONSE_YOU_HAVE_REGISTERED,
-    SERVER_RESPONSE_LOGIN_SUCCESS,
-    SERVER_RESPONSE_YOU_HAVE_LOGINED,
-    SERVER_RESPONSE_YOU_HAVE_NOT_LOGIN,
-    SERVER_RESPONSE_LOGIN_FAIL_UNREGISTERED_USERID,
-    SERVER_RESPONSE_LOGIN_FAIL_ERROR_PASSWORD,
-    SERVER_RESPONSE_LOGIN_FAIL_DUP_USERID,    // user id has been registered by other users
-    SERVER_RESPONSE_LOGIN_FAIL_SERVER_LIMITS, // server unable to handle more users
-    SERVER_RESPONSE_ALL_USERS_INFO,
-    SERVER_RESPONSE_ALL_FRIENDS_INFO,
-    SERVER_RESPONSE_LAUNCH_BATTLE_FAIL,
-    SERVER_RESPONSE_LAUNCH_BATTLE_SUCCESS,
-    SERVER_RESPONSE_YOURE_NOT_IN_BATTLE,
-    SERVER_RESPONSE_YOURE_ALREADY_IN_BATTLE,
-    SERVER_RESPONSE_INVITATION_SENT,
-    SERVER_RESPONSE_NOBODY_INVITE_YOU,
-    /* ----------------------------------------------- */
-    SERVER_MESSAGE,
-    SERVER_MESSAGE_DELIM,
-    SERVER_MESSAGE_FRIEND_LOGIN,
-    SERVER_MESSAGE_FRIEND_LOGOUT,
-    SERVER_MESSAGE_FRIEND_ACCEPT_BATTLE,
-    SERVER_MESSAGE_FRIEND_REJECT_BATTLE,
-    SERVER_MESSAGE_FRIEND_NOT_LOGIN,
-    SERVER_MESSAGE_FRIEND_ALREADY_IN_BATTLE,
-    SERVER_MESSAGE_INVITE_TO_BATTLE,
-    SERVER_MESSAGE_FRIEND_MESSAGE,
-    SERVER_MESSAGE_USER_QUIT_BATTLE,
-    SERVER_MESSAGE_BATTLE_DISBANDED,          // since no other users in this battle
-    SERVER_MESSAGE_BATTLE_INFORMATION,
-    SERVER_MESSAGE_BATTLE_PLAYER,
-    SERVER_MESSAGE_YOU_ARE_DEAD,
-    SERVER_MESSAGE_YOU_ARE_SHOOTED,
-    SERVER_MESSAGE_YOU_ARE_TRAPPED_IN_MAGMA,
-    SERVER_MESSAGE_YOU_GOT_BLOOD_VIAL,
-    SERVER_MESSAGE_YOU_GOT_MAGAZINE,
-    SERVER_MESSAGE_YOUR_MAGAZINE_IS_EMPTY,
-
-    SERVER_STATUS_QUIT,
-    SERVER_STATUS_FATAL,
-};
-
-
-enum {
-    DIR_UP,
-    DIR_DOWN,
-    DIR_LEFT,
-    DIR_RIGHT,
-    DIR_UP_LEFT,
-    DIR_UP_RIGHT,
-    DIR_DOWN_LEFT,
-    DIR_DOWN_RIGHT,
-};
-
-enum {
-    BATTLE_STATE_UNJOINED,
-    BATTLE_STATE_LIVE,
-    BATTLE_STATE_WITNESS,
-    BATTLE_STATE_DEAD,
-};
 #endif
